+++ CHANGELOG.rst
@@ -13,6 +13,7 @@ Added
 ~~~~~
 
 -  Custom user specific databases, use commands ``/save`` and ``/save_mode`` more information in ``/commands``
+-  Be able to show custom DB entries with ``/db_list``
 -  Add functionality to add alias commands just like a normal command but a string as ``command`` value, which points to
    a ``command_name``. Additionally ``title``, ``description``, ``hidden`` and ``group`` can be set.